fix: convert Fixed16 to float in rand01 functions for consistency
The rand01() functions were returning Fixed16 type when SCALAR_TYPE was FIXED16, which is inconsistent with the float return type expected by callers. This caused compilation errors in dependent code. Explicitly cast to float to ensure consistent return type across all SCALAR_TYPE configurations.

+++ include/math/MathUtil.h
@@ -242,7 +242,7 @@ inline Scalar rand01() {
         return static_cast<float>(r) / static_cast<float>(UINT32_MAX);
     } else {
         // Fixed16: use high 16 bits directly, no float conversion
-        return Fixed16::fromRaw(static_cast<int16_t>(r >> 16));
+        return static_cast<float>(Fixed16::fromRaw(static_cast<int16_t>(r >> 16)));
     }
 }
 
@@ -330,7 +330,7 @@ struct Random {
             return static_cast<float>(r) / static_cast<float>(UINT32_MAX);
         } else {
             // Fixed16: use high 16 bits directly, no float conversion
-            return Fixed16::fromRaw(static_cast<int16_t>(r >> 16));
+            return static_cast<float>(Fixed16::fromRaw(static_cast<int16_t>(r >> 16)));
         }
     }
